requirements.txt
tqdm
Pillow
torch
torchvision
transformers
sentencepiece
decord
dearpygui